+++ docs/sphinx/api/languages/python_api.rst
@@ -69,9 +69,15 @@ Data Types
     .. automethod:: __mul__
     .. automethod:: __rmul__
     .. automethod:: __getitem__
+    .. automethod:: slice
 
 .. autoclass:: cudaq::qubit
 .. autoclass:: cudaq::qreg
+
+.. autoclass:: cudaq::ComplexMatrix
+    .. automethod:: __getitem__ 
+    .. automethod:: minimal_eigenvalue
+    .. automethod:: __str__ 
 
 .. autoclass:: cudaq::SpinOperator
     :members:
@@ -84,6 +90,7 @@ Data Types
     .. automethod:: __mul__
     .. automethod:: __rmul__
     .. automethod:: __getitem__
+    .. automethod:: to_matrix
 
 .. autofunction:: cudaq::spin.i
 .. autofunction:: cudaq::spin.x
