+++ python/runtime/cudaq/builder/py_QuakeValue.cpp
@@ -54,10 +54,20 @@ void bindQuakeValue(py::module &mod) {
           "\nRaises:\n"
           "  RuntimeError: if `self` is a non-subscriptable "
           ":class:`QuakeValue`.\n")
-      .def("slice", &QuakeValue::slice,
-           "Return a slice of the given QuakeValue. The QuakeValue must be "
-           "list or qvec like. The first argument is the start index for the "
-           "slice, followed by the number of elements to extract.")
+      .def(
+          "slice", &QuakeValue::slice, py::arg("start"), py::arg("count"),
+          R"(Return a slice of the given :class:`QuakeValue` as a new :class:`QuakeValue`. 
+          
+               Note:
+                 The underlying :class:`QuakeValue` must be a `list` or `qvec`.
+                 
+               Args:
+                 start (int) : The index to begin the slice from.
+                 count (int) : The number of elements to extract after the `start` index.
+                   
+               Returns:
+                 :class:`QuakeValue` : A new `QuakeValue` containing a slice of `self` from
+                                         the `start`-th element to `start + count`-th element.)")
       /// @brief Bind the binary operators on `QuakeValue` class. Note:
       /// these are incompatible with the pybind11 built-in
       /// binary operators (see: cudaq::spin_op bindings). Instead,
